add role & more user functionality

+++ internal/models/requests/user_request.go
@@ -1,8 +1,8 @@
 package requests
 
 type SetUploadLimitRequest struct {
-	User  int `json:"user" validate:"required"`
-	Limit int `json:"limit" validate:"required"`
+	User  int   `json:"user" validate:"required"`
+	Limit int64 `json:"limit" validate:"required"`
 }
 
 type AddUserInRole struct {
